Add block-light channel via a uniform 32-bit voxel word
Voxel light was a single sky-only channel that merely scaled the PBR sun,
so at night (sun off) locally-lit voxels could not glow. Widen the packed
voxel word from 16-bit to 32-bit everywhere (resident memory, mesh worker,
wire, IndexedDB, server/LevelDB) — one consistent Uint32Array format — and
add an independent block-light channel at bit 16 for emitters (e.g. lava).

The low 16 bits stay byte-identical, so every existing pack/unpack shift and
mask is unchanged. Legacy 16-bit persisted records widen losslessly on read
(IndexedDB and LevelDB); light bits are recomputed client-side on ingest.

Lighting: sky light stays sky-column only (emitters no longer merged in);
block light is an emitter-seeded BFS into the new field, cleared and rebuilt
each relight so removing an emitter darkens correctly. LUT indices mask off
the new bits to keep the 2048-entry tables intact.

Shader: sky and block terms are combined by per-channel max, so the sun wins
in daylight (look unchanged) and block light wins at night / in caves. Block
light reaches the GPU as a new per-vertex blockLight attribute; colour and
intensity are live-tunable in the debug panel. New debug modes: SkyLight
(renamed from VoxelLight) and BlockLight.

+++ client/src/game/voxel/Chunk.ts
@@ -24,7 +24,7 @@ import {
  */
 export class Chunk extends ChunkData {
   /** Temporary data buffer for preview (not persisted) */
-  tempData: Uint16Array | null = null;
+  tempData: Uint32Array | null = null;
 
   /** Whether the chunk needs to be remeshed */
   dirty: boolean = true;
@@ -181,7 +181,7 @@ export class Chunk extends ChunkData {
 
   copyToTemp(): void {
     if (!this.tempData) {
-      this.tempData = new Uint16Array(VOXELS_PER_CHUNK);
+      this.tempData = new Uint32Array(VOXELS_PER_CHUNK);
     }
     this.tempData.set(this.data);
   }
